added gates

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -16,6 +16,11 @@ class AttendeeController extends Controller
 
     protected array $relations = ['user'];
 
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show']);
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -37,7 +42,7 @@ class AttendeeController extends Controller
     {
         $attendee = $this->loadRelationships(
             $event->attendees()->create([
-                'user_id' => 1
+                'user_id' => $request->user()->id
             ])
         );
 
@@ -58,7 +63,8 @@ class AttendeeController extends Controller
      * Remove the specified resource from storage.
      */
     public function destroy(Event $event, Attendee $attendee)
-    {
+    {   
+        $this->authorize('delete-attendee', [$event, $attendee]);
         $attendee->delete();
         return response(status: 204);
     }
